Actually fix packet loss
It turns out the thing I was missing was:
1. Updating to a newer ndppd
2. Having `ip` in its path so it can make calls to it

I confused myself as to what was and wasn't a working config because I
kept running ndppd outside of systemd to try and debug, and it would
work there... and then I'd flash the router config, and it'd be running
under systemd without `ip` and fail again. There's no error checking for
the system calls either, so it was a silent failure.

Once I realized that, this all made a lot more sense